Add function to disable auth checks on model

+++ src/AuthorizationRequired/AuthorizationRequired.php
@@ -8,6 +8,13 @@ use Gate;
 
 trait AuthorizationRequired
 {
+    public static $authorizationEnabled = true;
+
+    public static function disableAuthorization()
+    {
+        self::$authorizationEnabled = false;
+    }
+
     /**
      * Boot the authorization-required trait for a model.
      *
@@ -18,19 +25,19 @@ trait AuthorizationRequired
         static::addGlobalScope(new AuthorizationRequiredScope());
 
         self::creating(function ($self) {
-            if (!Gate::allows('create', $self)) {
+            if (self::$authorizationEnabled && !Gate::allows('create', $self)) {
                 throw new CreatePermissionException('Not allowed to create a ' . class_basename($self));
             }
         });
 
         self::updating(function ($self) {
-            if (!Gate::allows('update', $self)) {
+            if (self::$authorizationEnabled && !Gate::allows('update', $self)) {
                 throw new UpdatePermissionException('Not allowed to update this ' . class_basename($self));
             }
         });
 
         self::deleting(function ($self) {
-            if (!Gate::allows('delete', $self)) {
+            if (self::$authorizationEnabled && !Gate::allows('delete', $self)) {
                 throw new DeletePermissionException('Not allowed to delete this ' . class_basename($self));
             }
         });
